refactor: move around the caches
